Chore: move publish configures to root project
