+++ settings.gradle.kts
@@ -1,3 +1,3 @@
 include(":gradle-plugin")
-include("annotation")
-include("runtime")
+include(":annotation")
+include(":runtime")
